requirements.txt
streamlit>=1.28.0
weaviate-client>=3.25.0
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
numpy>=1.24.0
